correcting mob spell damage

+++ sql/char_effects.sql
@@ -8,7 +8,7 @@ SET SQL_MODE="NO_AUTO_VALUE_ON_ZERO";
 /*!40101 SET NAMES utf8 */;
 
 --
--- Base de données: `dspdb`
+-- Base de donnÃ©es: `dspdb`
 --
 
 -- --------------------------------------------------------
@@ -22,11 +22,11 @@ CREATE TABLE IF NOT EXISTS `char_effects` (
   `charid` int(10) unsigned NOT NULL,
   `effectid` smallint(5) unsigned NOT NULL,
   `icon` smallint(5) unsigned NOT NULL DEFAULT '0',
-  `power` smallint(5) unsigned NOT NULL DEFAULT '0',
+  `power` int unsigned NOT NULL DEFAULT '0',
   `tick` int(10) unsigned NOT NULL DEFAULT '0',
   `duration` int(10) unsigned NOT NULL DEFAULT '0',
   `subid` smallint(5) unsigned NOT NULL DEFAULT '0',
-  `subpower` smallint(5) NOT NULL DEFAULT '0',
+  `subpower` int NOT NULL DEFAULT '0',
   `tier` smallint(5) unsigned NOT NULL DEFAULT '0',
   KEY `charid` (`charid`)
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8;
